Update info command to allow for multiple repositories. (#1275)
For multiple repositories, the info command has been enhanced to search over all repositories when a backup set is not requested. If a backup set is requested, the --repo option is required if more than one repository is configured.

The json output from the command will now include information about each repo, including the status of the stanza on the repo. The text output has been updated to include the repo key to indicate where each backup is located.

A new status code of 'mixed' has been introduced for the overall stanza status and overall cipher type when they differ across repos. For the text output, when mixed is displayed, the status/cipher type will include additional lines indicating the status/cipher type for each repo.