Make remote storage objects writeable.
This wasn't exposed before because the remote protocol directly uses the storage driver, which bypasses the writeable checks.

However, the upcoming GCS driver explicitly requests write permissions so remote operations fail when a write is required.

It would be far better if the remote itself was marked as writeable but that will require much more work.