Convert storage option enums to StringId.
These enums map directly to option value strings so this removes some conditionals.

The conditionals were a poor choice to being with since they can't handle more than two values and make an assumption about the value that is not explicitly checked, which could lead to incorrect behavior when a new value is added.