Add temp mem context reset blocks in job processing loops.
Restore definitely needed to be doing cleanup, just as backup does. The archive-get, archive-push, and verify loop did not seem to be a significant source of leaks but that could change in the future so add resets.

Add temp mem context blocks in the job callbacks where they were missing.

Also switch to the prior context when creating a job, if possible, to save a move.